requirements.txt
streamlit
python-dotenv
openai
nest_asyncio
requests
PyPDF2
langchain
langchain-community
langchain-google-genai
faiss-cpu
streamlit
google-generativeai
requests
duckduckgo-search
Pillow
python-dotenv
asyncio-tools
aiohttp
beautifulsoup4
openai
stability-sdk
fastapi
uvicorn